Mesh: back in the root folder; moving it breaks MeshTests
