+++ src/main/kotlin/graphics/scenery/BoundingGrid.kt
@@ -1,6 +1,5 @@
 package graphics.scenery
 
-import graphics.scenery.primitives.Mesh
 import graphics.scenery.primitives.TextBoard
 import graphics.scenery.utils.extensions.*
 import org.joml.Vector3f
