+++ src/main/kotlin/graphics/scenery/Mesh.kt
@@ -1,4 +1,4 @@
-package graphics.scenery.primitives
+package graphics.scenery
 
 import org.joml.Vector3f
 import gnu.trove.map.hash.THashMap
@@ -6,6 +6,7 @@ import gnu.trove.set.hash.TLinkedHashSet
 import graphics.scenery.*
 import graphics.scenery.geometry.GeometryType
 import graphics.scenery.geometry.HasGeometry
+import graphics.scenery.primitives.PointCloud
 import graphics.scenery.textures.Texture
 import graphics.scenery.utils.Image
 import graphics.scenery.utils.LazyLogger
