+++ src/main/kotlin/graphics/scenery/controls/Hololens.kt
@@ -5,7 +5,7 @@ import org.joml.Vector3f
 import graphics.scenery.*
 import graphics.scenery.backends.Display
 import graphics.scenery.backends.vulkan.*
-import graphics.scenery.primitives.Mesh
+import graphics.scenery.Mesh
 import graphics.scenery.utils.LazyLogger
 import kotlinx.coroutines.*
 import org.joml.Quaternionf
