+++ src/main/kotlin/graphics/scenery/controls/TrackedStereoGlasses.kt
@@ -3,7 +3,7 @@ package graphics.scenery.controls
 import graphics.scenery.*
 import graphics.scenery.backends.Display
 import graphics.scenery.backends.vulkan.VulkanDevice
-import graphics.scenery.primitives.Mesh
+import graphics.scenery.Mesh
 import graphics.scenery.utils.LazyLogger
 import graphics.scenery.utils.extensions.plus
 import org.joml.*
