+++ src/main/kotlin/graphics/scenery/controls/behaviours/CreateCommand.kt
@@ -2,7 +2,7 @@ package graphics.scenery.controls.behaviours
 
 import graphics.scenery.Camera
 import graphics.scenery.Scene
-import graphics.scenery.primitives.Mesh
+import graphics.scenery.Mesh
 import graphics.scenery.backends.Renderer
 import graphics.scenery.utils.LazyLogger
 import org.joml.Vector2f
