+++ src/main/kotlin/graphics/scenery/geometry/Curve.kt
@@ -1,7 +1,7 @@
 package graphics.scenery.geometry
 
 import graphics.scenery.BufferUtils
-import graphics.scenery.primitives.Mesh
+import graphics.scenery.Mesh
 import graphics.scenery.utils.extensions.toFloatArray
 import org.joml.*
 import kotlin.Float.Companion.MIN_VALUE
