+++ src/main/kotlin/graphics/scenery/net/NodeSubscriber.kt
@@ -10,7 +10,7 @@ import graphics.scenery.geometry.GeometryType
 import graphics.scenery.primitives.Arrow
 import graphics.scenery.primitives.Cylinder
 import graphics.scenery.primitives.Line
-import graphics.scenery.primitives.Mesh
+import graphics.scenery.Mesh
 import graphics.scenery.utils.LazyLogger
 import graphics.scenery.utils.Statistics
 import graphics.scenery.volumes.TransferFunction
