+++ src/main/kotlin/graphics/scenery/primitives/Cone.kt
@@ -1,6 +1,7 @@
 package graphics.scenery.primitives
 
 import graphics.scenery.BufferUtils
+import graphics.scenery.Mesh
 import graphics.scenery.utils.extensions.minus
 import graphics.scenery.utils.extensions.plus
 import graphics.scenery.utils.extensions.times
