+++ src/main/kotlin/graphics/scenery/primitives/Plane.kt
@@ -1,6 +1,7 @@
 package graphics.scenery.primitives
 
 import graphics.scenery.BufferUtils
+import graphics.scenery.Mesh
 import org.joml.Vector3f
 
 /**
