+++ src/main/kotlin/graphics/scenery/proteins/Helix.kt
@@ -2,7 +2,7 @@ package graphics.scenery.proteins
 
 import graphics.scenery.geometry.Curve
 import graphics.scenery.geometry.Spline
-import graphics.scenery.primitives.Mesh
+import graphics.scenery.Mesh
 import org.joml.*
 
 /**
