+++ src/test/kotlin/graphics/scenery/tests/examples/advanced/BloodCellsExample.kt
@@ -3,7 +3,7 @@ package graphics.scenery.tests.examples.advanced
 import graphics.scenery.*
 import graphics.scenery.backends.Renderer
 import graphics.scenery.numerics.Random
-import graphics.scenery.primitives.Mesh
+import graphics.scenery.Mesh
 import graphics.scenery.utils.extensions.minus
 import org.joml.Vector3f
 import kotlin.concurrent.thread
