+++ src/test/kotlin/graphics/scenery/tests/examples/stresstests/PowerplantExample.kt
@@ -4,7 +4,7 @@ import org.joml.Vector3f
 import graphics.scenery.*
 import graphics.scenery.backends.Renderer
 import graphics.scenery.numerics.Random
-import graphics.scenery.primitives.Mesh
+import graphics.scenery.Mesh
 import kotlin.concurrent.thread
 
 /**
