+++ src/test/kotlin/graphics/scenery/tests/examples/stresstests/RungholtExample.kt
@@ -5,7 +5,7 @@ import graphics.scenery.*
 import graphics.scenery.backends.Renderer
 import graphics.scenery.controls.OpenVRHMD
 import graphics.scenery.numerics.Random
-import graphics.scenery.primitives.Mesh
+import graphics.scenery.Mesh
 import kotlin.concurrent.thread
 
 /**
