+++ src/test/kotlin/graphics/scenery/tests/unit/RainbowTests.kt
@@ -1,6 +1,6 @@
 package graphics.scenery.tests.unit
 
-import graphics.scenery.primitives.Mesh
+import graphics.scenery.Mesh
 import graphics.scenery.proteins.Rainbow
 import org.joml.Vector3f
 import org.junit.Test
